QueueITbackend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
fastapi==0.115.0
h11==0.16.0
httptools==0.7.1
idna==3.11
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
starlette==0.38.6
structlog==24.4.0
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
python-json-logger==2.0.7
sentry-sdk==2.19.0
prometheus-client==0.21.0
pytest==8.3.3
httpx==0.27.2
